+++ src/sas/qtgui/Plotting/Slicing/Interactors/ArcInteractor.py
