+++ src/sas/qtgui/Plotting/Slicing/Slicers/AnnulusSlicer.py
@@ -6,9 +6,9 @@ from sas.qtgui.Plotting.Plotter2D import Plotter2D
 import sas.qtgui.Utilities.GuiUtils as GuiUtils
 from sas.qtgui.Plotting.BaseInteractor import BaseInteractor
 from sas.qtgui.Plotting.PlotterData import Data1D
-from sas.qtgui.Plotting.Slicing.SlicerModel import SlicerModel
+from sas.qtgui.Plotting.Slicing.SlicerParameterWidget import SlicerParameterWidget
 
-class AnnulusInteractor(BaseInteractor[Plotter2D], SlicerModel):
+class AnnulusInteractor(BaseInteractor[Plotter2D], SlicerParameterWidget):
     """
     AnnulusInteractor plots a data1D average of an annulus area defined in a
     Data2D object. The data1D averaging itself is performed in sasdata by
@@ -21,7 +21,7 @@ class AnnulusInteractor(BaseInteractor[Plotter2D], SlicerModel):
     def __init__(self, base: Plotter2D, axes: Axes, item=None, color='black', zorder=3):
 
         BaseInteractor.__init__(self, base, axes, color=color)
-        SlicerModel.__init__(self)
+        SlicerParameterWidget.__init__(self)
 
         self.markers = []
         self.axes = axes
